Seperated api and server calls for security and ensure that sqids generates only numbers.
